Finish startup hooks
