+++ pkg/config/config.go
@@ -29,6 +29,9 @@ import (
 	"time"
 
 	"github.com/pbnjay/memory"
+	"github.com/siglens/siglens/pkg/config/common"
+	commonconfig "github.com/siglens/siglens/pkg/config/common"
+	"github.com/siglens/siglens/pkg/hooks"
 	segutils "github.com/siglens/siglens/pkg/segment/utils"
 	"github.com/siglens/siglens/pkg/utils"
 	log "github.com/sirupsen/logrus"
@@ -41,119 +44,7 @@ const RunModFilePath = "data/common/runmod.cfg"
 
 var configFileLastModified uint64
 
-type DeploymentType uint8
-
-const (
-	SingleNode = iota + 1
-	SingleNodeS3
-	DistributedS3
-)
-
-func (d DeploymentType) String() string {
-	return [...]string{"INVALID", "SingleNode", "SingleNodeS3", "DistributedS3"}[d]
-}
-
-type S3Config struct {
-	Enabled      bool   `yaml:"enabled"`
-	BucketName   string `yaml:"bucketName"`
-	BucketPrefix string `yaml:"bucketPrefix"`
-	RegionName   string `yaml:"regionName"`
-}
-
-type EtcdConfig struct {
-	Enabled  bool     `yaml:"enabled"`
-	SeedUrls []string `yaml:"seedUrls"`
-}
-
-type EmailConfig struct {
-	SmtpHost         string `yaml:"smtpHost"`
-	SmtpPort         int    `yaml:"smtpPort"`
-	SenderEmail      string `yaml:"senderEmail"`
-	GmailAppPassword string `yaml:"gmailAppPassword"`
-}
-
-type LogConfig struct {
-	LogPrefix             string `yaml:"logPrefix"`             // Prefix of log file. Can be a directory. if empty will log to stdout
-	LogFileRotationSizeMB int    `yaml:"logFileRotationSizeMB"` //Max size of log file in megabytes
-	CompressLogFile       bool   `yaml:"compressLogFile"`
-}
-
-type TLSConfig struct {
-	Enabled    bool   `yaml:"enabled"`
-	ACMEFolder string `yaml:"acmeFolder"` // folder to store acme certificates
-}
-
-type AlertConfig struct {
-	Enabled  bool   `yaml:"enabled"`
-	Provider string `yaml:"provider"`
-	Host     string `yaml:"host"`
-	Port     uint64 `yaml:"port"`
-	User     string `yaml:"user"`
-	Password string `yaml:"password"`
-	Dbname   string `yaml:"dbname"`
-}
-
-type DatabaseConfig struct {
-	Enabled  bool   `yaml:"enabled"`
-	Provider string `yaml:"provider"`
-	Host     string `yaml:"host"`
-	Port     uint64 `yaml:"port"`
-	User     string `yaml:"user"`
-	Password string `yaml:"password"`
-	Dbname   string `yaml:"dbname"`
-}
-
-/*  If you add a new config parameters to the Configuration struct below, make sure to add the default value
-assignment in the following functions
-1) ExtractConfigData function
-2) InitializeDefaultConfig function */
-
-// If you add a new config parameters to the Configuration struct below, make sure to add a descriptive info in server.yaml
-type Configuration struct {
-	IngestPort                 uint64   `yaml:"ingestPort"`           // Port for ingestion server
-	QueryPort                  uint64   `yaml:"queryPort"`            // Port used for query server
-	PsqlPort                   uint64   `yaml:"psqlPort"`             // Port used for sql server
-	EventTypeKeywords          []string `yaml:"eventTypeKeywords"`    //Required event type keyword
-	QueryNode                  string   `yaml:"queryNode"`            //Node to enable/disable all query endpoints
-	IngestNode                 string   `yaml:"ingestNode"`           //Node to enable/disable all ingest endpoints
-	SegFlushIntervalSecs       int      `yaml:"segFlushIntervalSecs"` // Time Interval after which to write to segfile
-	DataPath                   string   `yaml:"dataPath"`
-	RetentionHours             int      `yaml:"retentionHours"`
-	TimeStampKey               string   `yaml:"timestampKey"`
-	MaxSegFileSize             uint64   `yaml:"maxSegFileSize"` // segment file size (in bytes)
-	LicenseKeyPath             string   `yaml:"licenseKeyPath"`
-	ESVersion                  string   `yaml:"esVersion"`
-	Debug                      bool     `yaml:"debug"`                  // debug logging
-	MemoryThresholdPercent     uint64   `yaml:"memoryThresholdPercent"` // percent of all available free data allocated for loading micro indices in memory
-	DataDiskThresholdPercent   uint64   `yaml:"dataDiskThresholdPercent"`
-	GRPCPort                   uint64   `yaml:"grpcPort"` // Address to listen for GRPC connections
-	S3IngestQueueName          string   `yaml:"s3IngestQueueName"`
-	S3IngestQueueRegion        string   `yaml:"s3IngestQueueRegion"`
-	S3IngestBufferSize         uint64   `yaml:"s3IngestBufferSize"`
-	MaxParallelS3IngestBuffers uint64   `yaml:"maxParallelS3IngestBuffers"`
-	SSInstanceName             string   `yaml:"ssInstanceName"`
-	PQSEnabled                 string   `yaml:"pqsEnabled"` // is pqs enabled?
-	pqsEnabledConverted        bool     // converted bool value of PQSEnabled yaml
-	SafeServerStart            bool     `yaml:"safeMode"`         // if set to true, siglens will start a mock webserver with a custom health handler. Actual server will NOT be started
-	AnalyticsEnabled           string   `yaml:"analyticsEnabled"` // is analytics enabled?
-	analyticsEnabledConverted  bool
-	AgileAggsEnabled           string `yaml:"agileAggsEnabled"` // should we read/write AgileAggsTrees?
-	AgileAggsEnabledConverted  bool
-	QueryHostname              string         `yaml:"queryHostname"` // hostname of the query server. i.e. if DNS is https://cloud.siglens.com, this should be cloud.siglens.com
-	IngestUrl                  string         `yaml:"ingestUrl"`     // full address of the ingest server, including scheme and port, e.g. https://ingest.siglens.com:8080
-	S3                         S3Config       `yaml:"s3"`            // s3 related config
-	Etcd                       EtcdConfig     `yaml:"etcd"`          // Etcd related config
-	Log                        LogConfig      `yaml:"log"`           // Log related config
-	TLS                        TLSConfig      `yaml:"tls"`           // TLS related config
-	EmailConfig                EmailConfig    `yaml:"emailConfig"`
-	DatabaseConfig             DatabaseConfig `yaml:"minionSearch"`
-}
-
-type RunModConfig struct {
-	PQSEnabled bool `json:"pqsEnabled"`
-}
-
-var runningConfig Configuration
+var runningConfig common.Configuration
 var configFilePath string
 
 var parallelism int64
@@ -198,7 +89,7 @@ func GetDataDiskThresholdPercent() uint64 {
 	return runningConfig.DataDiskThresholdPercent
 }
 
-func GetRunningConfig() *Configuration {
+func GetRunningConfig() *common.Configuration {
 	return &runningConfig
 }
 
@@ -294,13 +185,20 @@ func GetQueryHostname() string {
 	return runningConfig.QueryHostname
 }
 
-func GetEtcdConfig() EtcdConfig {
+func GetEtcdConfig() common.EtcdConfig {
 	return runningConfig.Etcd
 }
 
 // returns SmtpHost, SmtpPort, SenderEmail and GmailAppPassword
 func GetEmailConfig() (string, int, string, string) {
 	return runningConfig.EmailConfig.SmtpHost, runningConfig.EmailConfig.SmtpPort, runningConfig.EmailConfig.SenderEmail, runningConfig.EmailConfig.GmailAppPassword
+}
+
+func SetEmailConfig(smtpHost string, smtpPort int, senderEmail string, gmailAppPassword string) {
+	runningConfig.EmailConfig.SmtpHost = smtpHost
+	runningConfig.EmailConfig.SmtpPort = smtpPort
+	runningConfig.EmailConfig.SenderEmail = senderEmail
+	runningConfig.EmailConfig.GmailAppPassword = gmailAppPassword
 }
 
 func GetUIDomain() string {
@@ -325,7 +223,7 @@ func IsDebugMode() bool {
 }
 
 func IsPQSEnabled() bool {
-	return runningConfig.pqsEnabledConverted
+	return runningConfig.PQSEnabledConverted
 }
 
 func IsAggregationsEnabled() bool {
@@ -338,7 +236,7 @@ func SetAggregationsFlag(enabled bool) {
 }
 
 func IsAnalyticsEnabled() bool {
-	return runningConfig.analyticsEnabledConverted
+	return runningConfig.AnalyticsEnabledConverted
 }
 
 func IsSafeMode() bool {
@@ -469,7 +367,7 @@ func SetMaxParallelS3IngestBuffers(maxBuf uint64) {
 	runningConfig.MaxParallelS3IngestBuffers = maxBuf
 }
 func SetPQSEnabled(enabled bool) {
-	runningConfig.pqsEnabledConverted = enabled
+	runningConfig.PQSEnabledConverted = enabled
 	runningConfig.PQSEnabled = strconv.FormatBool(enabled)
 }
 
@@ -481,19 +379,19 @@ func IsMultinodeEnabled() bool {
 	return runningConfig.Etcd.Enabled
 }
 
-func ValidateDeployment() (DeploymentType, error) {
+func ValidateDeployment() (commonconfig.DeploymentType, error) {
 
 	if runningConfig.Etcd.Enabled {
 		if runningConfig.S3.Enabled {
-			return DistributedS3, nil
+			return commonconfig.DistributedS3, nil
 		}
 		return 0, fmt.Errorf("etcd must be enabled with S3")
 	}
 	if IsQueryNode() && IsIngestNode() {
 		if runningConfig.S3.Enabled {
-			return SingleNodeS3, nil
+			return commonconfig.SingleNodeS3, nil
 		}
-		return SingleNode, nil
+		return commonconfig.SingleNode, nil
 	}
 	return 0, fmt.Errorf("single node deployment must have both query and ingest in the same node")
 }
@@ -521,7 +419,7 @@ func InitConfigurationData() error {
 		return err
 	}
 	runningConfig = config
-	var readConfig RunModConfig
+	var readConfig common.RunModConfig
 	readConfig, err = ReadRunModConfig(RunModFilePath)
 	if err != nil && !os.IsNotExist(err) {
 		log.Errorf("InitConfigurationData: Failed to read runmod config: %v, config: %+v", err, readConfig)
@@ -551,7 +449,7 @@ func InitializeDefaultConfig() {
 	// 3. And Here.
 	// ************************************
 
-	runningConfig = Configuration{
+	runningConfig = common.Configuration{
 		IngestPort:                 8081,
 		QueryPort:                  5122,
 		IngestUrl:                  "",
@@ -560,7 +458,7 @@ func InitializeDefaultConfig() {
 		IngestNode:                 "true",
 		SegFlushIntervalSecs:       5,
 		DataPath:                   "data/",
-		S3:                         S3Config{false, "", "", ""},
+		S3:                         common.S3Config{false, "", "", ""},
 		RetentionHours:             24 * 90,
 		TimeStampKey:               "timestamp",
 		MaxSegFileSize:             1_073_741_824,
@@ -576,19 +474,19 @@ func InitializeDefaultConfig() {
 		MaxParallelS3IngestBuffers: 10,
 		SSInstanceName:             "",
 		PQSEnabled:                 "false",
-		pqsEnabledConverted:        false,
+		PQSEnabledConverted:        false,
 		SafeServerStart:            false,
 		AnalyticsEnabled:           "false",
-		analyticsEnabledConverted:  false,
+		AnalyticsEnabledConverted:  false,
 		AgileAggsEnabled:           "true",
 		AgileAggsEnabledConverted:  true,
 		QueryHostname:              "",
-		Log:                        LogConfig{"", 100, false},
-		TLS:                        TLSConfig{false, "certs/"},
-		DatabaseConfig:             DatabaseConfig{Enabled: true, Provider: "sqlite"},
+		Log:                        common.LogConfig{"", 100, false},
+		TLS:                        common.TLSConfig{false, "certs/"},
+		DatabaseConfig:             common.DatabaseConfig{Enabled: true, Provider: "sqlite"},
 	}
 	_ = InitDerivedConfig("test-uuid") // This is only used for testing
-	runningConfig.EmailConfig = EmailConfig{"smtp.gmail.com", 587, "doe1024john@gmail.com", " "}
+	runningConfig.EmailConfig = common.EmailConfig{"smtp.gmail.com", 587, "doe1024john@gmail.com", " "}
 }
 
 func InitializeTestingConfig() {
@@ -597,14 +495,14 @@ func InitializeTestingConfig() {
 	SetDataPath("data/")
 }
 
-func ReadRunModConfig(fileName string) (RunModConfig, error) {
+func ReadRunModConfig(fileName string) (common.RunModConfig, error) {
 	_, err := os.Stat(fileName)
 	if os.IsNotExist(err) {
 		log.Infof("ReadRunModConfig:Config file '%s' does not exist. Awaiting user action to create it.", fileName)
-		return RunModConfig{}, err
+		return common.RunModConfig{}, err
 	} else if err != nil {
 		log.Errorf("ReadRunModConfig:Error accessing config file '%s': %v", fileName, err)
-		return RunModConfig{}, err
+		return common.RunModConfig{}, err
 	}
 
 	jsonData, err := os.ReadFile(fileName)
@@ -614,8 +512,8 @@ func ReadRunModConfig(fileName string) (RunModConfig, error) {
 	return ExtractReadRunModConfig(jsonData)
 }
 
-func ExtractReadRunModConfig(jsonData []byte) (RunModConfig, error) {
-	var runModConfig RunModConfig
+func ExtractReadRunModConfig(jsonData []byte) (common.RunModConfig, error) {
+	var runModConfig common.RunModConfig
 	err := json.Unmarshal(jsonData, &runModConfig)
 	if err != nil {
 		log.Errorf("ExtractReadRunModConfig:Failed to parse runmod.cfg: %v", err)
@@ -626,16 +524,21 @@ func ExtractReadRunModConfig(jsonData []byte) (RunModConfig, error) {
 	return runModConfig, nil
 }
 
-func ReadConfigFile(fileName string) (Configuration, error) {
+func ReadConfigFile(fileName string) (common.Configuration, error) {
 	yamlData, err := os.ReadFile(fileName)
 	if err != nil {
 		log.Errorf("Cannot read input fileName = %v, err=%v", fileName, err)
 	}
-	return ExtractConfigData(yamlData)
+
+	if hook := hooks.GlobalHooks.ExtractConfigHook; hook != nil {
+		return hook(yamlData)
+	} else {
+		return ExtractConfigData(yamlData)
+	}
 }
 
-func ExtractConfigData(yamlData []byte) (Configuration, error) {
-	var config Configuration
+func ExtractConfigData(yamlData []byte) (common.Configuration, error) {
+	var config common.Configuration
 	err := yaml.Unmarshal(yamlData, &config)
 	if err != nil {
 		log.Errorf("Error parsing yaml err=%v", err)
@@ -676,7 +579,7 @@ func ExtractConfigData(yamlData []byte) (Configuration, error) {
 		pqsEnabled = false
 		config.PQSEnabled = "false"
 	}
-	config.pqsEnabledConverted = pqsEnabled
+	config.PQSEnabledConverted = pqsEnabled
 
 	if len(config.AnalyticsEnabled) <= 0 {
 		config.AnalyticsEnabled = "true"
@@ -687,7 +590,7 @@ func ExtractConfigData(yamlData []byte) (Configuration, error) {
 		analyticsEnabled = true
 		config.AnalyticsEnabled = "true"
 	}
-	config.analyticsEnabledConverted = analyticsEnabled
+	config.AnalyticsEnabledConverted = analyticsEnabled
 
 	if len(config.AgileAggsEnabled) <= 0 {
 		config.AgileAggsEnabled = "true"
@@ -783,7 +686,7 @@ func ExtractConfigData(yamlData []byte) (Configuration, error) {
 	return config, nil
 }
 
-func SetConfig(config Configuration) {
+func SetConfig(config common.Configuration) {
 	runningConfig = config
 }
